requirements.txt
streamlit>=1.38.0
Pillow>=10.0.0